feat: Add initial inventory update for template
Update the template so it display an update button which can be used to update
individual inventories in the future. Need to style the button so it and future
buttons are centred correct. Also need to change color / style of some of the
buttons to make it more obvious, perhaps adding icons as well